PostToS3/requirements.txt
blinker==1.8.2
boto3==1.34.161
botocore==1.34.161
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==8.2.0
inspirational_quotes==0.0.11
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
ujson==5.10.0
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.20.0
